Adds text names and respawns.
When the player connects to the server, they will be prompted for a
name (which is optional). The server then spawns the player in
a random spot. After dying, the websocket connection is now preserved
(instead of instantly DCing which was a hack). It will just not
draw your player or let them interact but you will be looking at
the most recent location you were in before getting eaten. You
will be prompted to hit play again, optionally changing the name
provided.