Adds speed variance and maximum radius.
I've changed a bunch of config constants here, like the board is
larger and square. There is a maximum radius, and speed slows
down linearly as you go from the minimum to the maximum radius.
Additionally once you're over a shrink radius threshold, you
will start to shrink back down to it so that players don't stay
at the max radius forever.